+++ CHANGES.rst
@@ -1,3 +1,19 @@
+Release 9.1.1 (in development)
+==============================
+
+Dependencies
+------------
+
+* #14026: Added `regex >= 2025.7.29`_, to support parsing for Unicode classes.
+
+  .. _regex >= 2025.7.29: _https://github.com/mrabarnett/mrab-regex
+
+Features added
+--------------
+
+Bugs fixed
+----------
+
 Release 9.1.0 (released Dec 31, 2025)
 =====================================